make zonal diagnostic for all kx

+++ STELLA_CODE/read_namelists_from_input_file/namelist_diagnostics.f90
@@ -737,7 +737,7 @@ contains
          implicit none
          integer :: ios
          integer, parameter :: max_kxs = 100
-         real(kind=8) :: zonal_kxs(max_kxs)
+         real :: zonal_kxs(max_kxs)
          integer :: j
 
          namelist /diagnostics_zonal/ write_g_vs_zvpas_zonal, write_free_energy, &
